feat(Socket): (Re)connection strategy
Totally new implementation of the sokcet's connectionstrategy. Basically, we have removed all attempts of calculating advanced combinations of the state between the server side and the client(s). The reconnection and retry strategy is simpler and more robust. This, in combination with a more efficient command timeout handling, gives better results in the event of server restart or shutdown.

It also removes wrongly raised error-events for missed conenction attempts.

Resolves #66