feat(Version): Added version promise to CasparCG
added getCasparCGVersion(refresh: boolean): Promise<CasparCGVersion> to CasparCG. This follows the pattern of info- and paths-promises. This also removes the need for autoServerVersion which was a bad implementation with race conditions. Version commands are now done just in time, however: we need a way of putting this kind of commands at the top of the queue when chained with other commands (INFO CONFIG).

Resolves #73